Updated Validation Logic
