+++ sdk/entra/Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents/src/Framework/ActionableCloudEventResponse.cs
@@ -35,16 +35,5 @@ namespace Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents.Framework
 
             base.Invalidate();
         }
-
-        /// <summary>
-        /// If the response has no actions, mark as failed.
-        /// </summary>
-        internal override void ValidateActions()
-        {
-            if (Actions == null || Actions.Count == 0 || Actions[0] == null)
-            {
-                MarkAsFailed(new Exception($"TokenIssuanceStartResponse: {AuthenticationEventResource.Ex_No_Action}"), false);
-            }
-        }
     }
 }
