+++ sdk/entra/Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents/tests/TestAuthResponse.cs
@@ -5,23 +5,20 @@ using System.Net;
 namespace Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents.Tests
 {
 
-	internal class TestAuthResponse : AuthenticationEventResponse
-	{
-		internal TestAuthResponse(HttpStatusCode code, string content)
-		: this(code)
-		{
-			Content = new StringContent(content);
-		}
+    internal class TestAuthResponse : AuthenticationEventResponse
+    {
+        internal TestAuthResponse(HttpStatusCode code, string content)
+        : this(code)
+        {
+            Content = new StringContent(content);
+        }
 
-		internal TestAuthResponse(HttpStatusCode code)
-		{
-			StatusCode = code;
-		}
+        internal TestAuthResponse(HttpStatusCode code)
+        {
+            StatusCode = code;
+        }
 
-		internal override void Invalidate()
-		{ }
-
-		internal override void ValidateActions()
-		{ }
-	}
+        internal override void Invalidate()
+        { }
+    }
 }
