+++ sdk/entra/Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents/tests/TestHelper.cs
@@ -1,5 +1,6 @@
 ﻿//using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents.TokenIssuanceStart;
+using Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents.TokenIssuanceStart.Actions;
 using Microsoft.Azure.WebJobs.Host.Executors;
 using Microsoft.Extensions.Logging;
 using Moq;
@@ -320,6 +321,71 @@ namespace Microsoft.Azure.WebJobs.Extensions.AuthenticationEvents.Tests
             input = input.Trim();
             return (input.StartsWith("{", StringComparison.OrdinalIgnoreCase) && input.EndsWith("}", StringComparison.OrdinalIgnoreCase))
                 || (input.StartsWith("[", StringComparison.OrdinalIgnoreCase) && input.EndsWith("]", StringComparison.OrdinalIgnoreCase));
+        }
+
+        public enum ActionTestTypes
+        {
+            NullClaims,
+            EmptyClaims,
+            NullClaimId,
+            EmptyClaimsId,
+            EmptyValueString,
+            NullValue,
+            EmptyValueArray,
+            EmptyValueStringArray,
+            EmptyMixedArray,
+            NullAction
+        }
+
+        public static (TokenIssuanceAction action, HttpStatusCode expectReturnCode, string expectedResponse) GetActionTestExepected(ActionTestTypes actionTestTypes)
+        {
+            switch (actionTestTypes)
+            {
+                case ActionTestTypes.NullClaims:
+                    return (new ProvideClaimsForToken(null),
+                        HttpStatusCode.BadRequest,
+                        "{\"errors\":[\"TokenIssuanceStartResponse: ProvideClaimsForToken: The Claims field is required.\"]}");
+                case ActionTestTypes.EmptyClaims:
+                    return (new ProvideClaimsForToken(),
+                        HttpStatusCode.OK,
+                        "{\"data\":{\"@odata.type\":\"microsoft.graph.onTokenIssuanceStartResponseData\",\"actions\":[{\"@odata.type\":\"microsoft.graph.tokenIssuanceStart.provideClaimsForToken\",\"claims\":{}}]}}");
+                case ActionTestTypes.NullClaimId:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim(null, string.Empty) }),
+                    HttpStatusCode.BadRequest,
+                    "{\"errors\":[\"TokenIssuanceStartResponse: ProvideClaimsForToken: TokenClaim: The Id field is required.\"]}");
+                case ActionTestTypes.EmptyClaimsId:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim(String.Empty, string.Empty) }),
+                    HttpStatusCode.BadRequest,
+                    "{\"errors\":[\"TokenIssuanceStartResponse: ProvideClaimsForToken: TokenClaim: The Id field is required.\"]}");
+                case ActionTestTypes.EmptyValueString:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim("key", string.Empty) }),
+                    HttpStatusCode.OK,
+                    "{\"data\":{\"@odata.type\":\"microsoft.graph.onTokenIssuanceStartResponseData\",\"actions\":[{\"@odata.type\":\"microsoft.graph.tokenIssuanceStart.provideClaimsForToken\",\"claims\":{\"key\":\"\"}}]}}");
+                case ActionTestTypes.NullValue:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim("key", null) }),
+                    HttpStatusCode.OK,
+                    "{\"data\":{\"@odata.type\":\"microsoft.graph.onTokenIssuanceStartResponseData\",\"actions\":[{\"@odata.type\":\"microsoft.graph.tokenIssuanceStart.provideClaimsForToken\",\"claims\":{\"key\":null}}]}}");
+                case ActionTestTypes.EmptyValueArray:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim("key", new string[] { }) }),
+                    HttpStatusCode.OK,
+                    "{\"data\":{\"@odata.type\":\"microsoft.graph.onTokenIssuanceStartResponseData\",\"actions\":[{\"@odata.type\":\"microsoft.graph.tokenIssuanceStart.provideClaimsForToken\",\"claims\":{\"key\":[]}}]}}");
+                case ActionTestTypes.EmptyValueStringArray:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim("key", new string[] { String.Empty, String.Empty }) }),
+                    HttpStatusCode.OK,
+                    "{\"data\":{\"@odata.type\":\"microsoft.graph.onTokenIssuanceStartResponseData\",\"actions\":[{\"@odata.type\":\"microsoft.graph.tokenIssuanceStart.provideClaimsForToken\",\"claims\":{\"key\":[\"\",\"\"]}}]}}");
+                case ActionTestTypes.EmptyMixedArray:
+                    return (new ProvideClaimsForToken(new TokenClaim[] { new TokenClaim("key", new string[] { String.Empty, null, " " }) }),
+                    HttpStatusCode.OK,
+                    "{\"data\":{\"@odata.type\":\"microsoft.graph.onTokenIssuanceStartResponseData\",\"actions\":[{\"@odata.type\":\"microsoft.graph.tokenIssuanceStart.provideClaimsForToken\",\"claims\":{\"key\":[\"\",null,\" \"]}}]}}");
+                case ActionTestTypes.NullAction:
+                    return (null,
+                    HttpStatusCode.BadRequest,
+                    "{\"errors\":[\"TokenIssuanceStartResponse: No Actions Found. Please supply atleast one action.\"]}");
+                default:
+                    return (null,
+                    HttpStatusCode.InternalServerError,
+                    null);
+            }
         }
     }
 }
